Move Groovy and Kotlin tests of recipes to rewrite-groovy/ and rewrite-kotlin/
