+++ rewrite-groovy/src/test/java/org/openrewrite/groovy/format/NoWhitespaceBeforeTest.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2021 the original author or authors.
+ * <p>
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * <p>
+ * https://www.apache.org/licenses/LICENSE-2.0
+ * <p>
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.openrewrite.groovy.format;
+
+import org.junit.jupiter.api.Test;
+import org.openrewrite.Issue;
+import org.openrewrite.java.format.NoWhitespaceBefore;
+import org.openrewrite.test.RecipeSpec;
+import org.openrewrite.test.RewriteTest;
+
+import static org.openrewrite.groovy.Assertions.groovy;
+
+class NoWhitespaceBeforeTest implements RewriteTest {
+    @Override
+    public void defaults(RecipeSpec spec) {
+        spec.recipe(new NoWhitespaceBefore());
+    }
+
+    @Issue("https://github.com/openrewrite/rewrite-static-analysis/issues/233")
+    @Test
+    void jenkinsLibrary() {
+        rewriteRun(groovy("library 'someLibrary@version'"));
+    }
+}
