+++ rewrite-java-test/build.gradle.kts
@@ -3,7 +3,6 @@ plugins {
 }
 
 recipeDependencies {
-    parserClasspath("jakarta.persistence:jakarta.persistence-api:3.1.0")
     testParserClasspath("jakarta.validation:jakarta.validation-api:3.0.2")
     testParserClasspath("javax.validation:validation-api:1.1.0.Final")
     testParserClasspath("org.hibernate:hibernate-validator:5.4.3.Final")
@@ -14,12 +13,9 @@ dependencies {
     implementation(project(":rewrite-java"))
     implementation(project(":rewrite-test"))
 
-    testImplementation(project(":rewrite-groovy"))
-    testImplementation(project(":rewrite-kotlin"))
     testImplementation("io.github.classgraph:classgraph:latest.release")
     testImplementation("org.junit-pioneer:junit-pioneer:2.0.0")
     testRuntimeOnly(project(":rewrite-java-21"))
-    testRuntimeOnly("jakarta.persistence:jakarta.persistence-api:3.1.0")
     testRuntimeOnly("org.apache.hbase:hbase-shaded-client:2.4.11")
     testRuntimeOnly("com.google.guava:guava:latest.release")
     testRuntimeOnly("org.mapstruct:mapstruct:latest.release")
