+++ rewrite-java-test/src/test/java/org/openrewrite/java/FindMissingTypesTest.java
@@ -17,18 +17,15 @@ package org.openrewrite.java;
 
 import org.junit.jupiter.api.Test;
 import org.openrewrite.DocumentExample;
-import org.openrewrite.InMemoryExecutionContext;
 import org.openrewrite.Issue;
 import org.openrewrite.java.search.FindMissingTypes;
 import org.openrewrite.java.tree.JavaType;
-import org.openrewrite.kotlin.KotlinParser;
 import org.openrewrite.test.RecipeSpec;
 import org.openrewrite.test.RewriteTest;
 import org.openrewrite.test.TypeValidation;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.openrewrite.java.Assertions.java;
-import static org.openrewrite.kotlin.Assertions.kotlin;
 
 @SuppressWarnings("ClassInitializerMayBeStatic")
 class FindMissingTypesTest implements RewriteTest {
@@ -269,59 +266,4 @@ class FindMissingTypesTest implements RewriteTest {
         );
     }
 
-    @Test
-    void kotlinClassReference() {
-        rewriteRun(
-          spec -> spec.parser(
-            KotlinParser.builder().classpathFromResources(new InMemoryExecutionContext(), "jakarta.persistence-api")),
-          kotlin(
-            """
-              package com.some.other
-
-              class EventPublisher {
-              }
-              """
-          ),
-          kotlin(
-            """
-              package com.some.card
-
-              import com.some.other.EventPublisher
-              import jakarta.persistence.EntityListeners
-
-              @EntityListeners(EventPublisher::class)
-              data class Card(
-              )
-              """
-          )
-        );
-    }
-
-    @Test
-    void invalidKotlinClassReference() {
-        rewriteRun(
-          spec -> spec.parser(
-            KotlinParser.builder().classpathFromResources(new InMemoryExecutionContext(), "jakarta.persistence-api")),
-          kotlin(
-            """
-              package com.some.card
-
-              import jakarta.persistence.EntityListeners
-
-              @EntityListeners(EventPublisher::class)
-              data class Card(
-              )
-              """,
-            """
-              package com.some.card
-
-              import jakarta.persistence.EntityListeners
-
-              @EntityListeners(/*~~(MemberReference Parameterized type is missing or malformed)~~>*//*~~(Identifier type is missing or malformed)~~>*/EventPublisher::class)
-              data class Card(
-              )
-              """
-          )
-        );
-    }
 }
