+++ rewrite-java-test/src/test/java/org/openrewrite/java/format/NoWhitespaceBeforeTest.java
@@ -21,7 +21,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.DisabledOnOs;
 import org.junit.jupiter.api.condition.OS;
 import org.openrewrite.DocumentExample;
-import org.openrewrite.Issue;
 import org.openrewrite.Tree;
 import org.openrewrite.java.JavaParser;
 import org.openrewrite.java.style.Checkstyle;
@@ -39,7 +38,6 @@ import java.util.function.UnaryOperator;
 
 import static java.util.Collections.emptySet;
 import static java.util.Collections.singletonList;
-import static org.openrewrite.groovy.Assertions.groovy;
 import static org.openrewrite.java.Assertions.java;
 
 @SuppressWarnings({
@@ -753,12 +751,6 @@ class NoWhitespaceBeforeTest implements RewriteTest {
               """
           )
         );
-    }
-
-    @Issue("https://github.com/openrewrite/rewrite-static-analysis/issues/233")
-    @Test
-    void jenkinsLibrary() {
-        rewriteRun(groovy("library 'someLibrary@version'"));
     }
 
     private static Consumer<SourceSpec<J.CompilationUnit>> autoFormatIsIdempotent() {
